Adding new class to simplify PTP keys management
Each contact will now have a "latchword" property instead of a "key" property. With this change I am centralizing every single point-to-point encryption function into one contact-decoupled class. This was necessary since the Contact class has been growing exponentially due to new incoming fancy functionalities. Hence, decoupling the security architecture is handy in order to keep the code as clean as it was intended since the very first beggining of this project.